+++ README.md
@@ -29,3 +29,21 @@ This template was initially based on [Jason Tyler's Template](https://github.com
 6. If you need to uninstall the template
     - Open command prompt and set the current folder to the root of the repo
     - Execute 'dotnet new uninstall .' command in the command prompt
+
+## Authentication
+
+KeyCloak is setup as IDP with some defaults:
+
+Groups
+- Admin
+- Standard
+
+Roles
+- standard-user-role
+  - Can Create, Update and View
+- admin-user-role
+  - Can Delete
+  
+Users (Password is password)
+- AdminUser
+- StandardUser
