+++ src/MinimalApiTemplate.Api/Common/Constants.cs
@@ -20,18 +20,5 @@ public static class Constants
         public const string CorrelationId = "x-correlation-id";
         public const string UserProfileId = "UserProfileId";
         public const string Authorization = "Authorization";
-    }
-
-    public static class Roles
-    {
-        public const string StandardUserRole = "standard-user-role";
-        public const string AdminUserRole = "admin-user-role";
-    }
-
-    public static class Policies
-    {
-        public const string StandardUser = "StandardUser";
-        public const string AdminUser = "AdminUser";
-    }
-   
+    } 
 }
