+++ src/MinimalApiTemplate.Api/Endpoints/V1/TodoItems/GetTodoItemsWithPagination/GetTodoItemsWithPaginationEndpoint.cs
@@ -1,6 +1,7 @@
 ﻿using MinimalApiTemplate.Api.Common.Models;
 using MinimalApiTemplate.Application.Features.TodoItems.Queries.GetTodoItemsWithPagination;
 using static MinimalApiTemplate.Api.Common.Constants;
+using static MinimalApiTemplate.Application.Common.Constants;
 
 namespace MinimalApiTemplate.Api.Endpoints.V1.TodoItems.GetTodoItemsWithPagination;
