+++ src/MinimalApiTemplate.Application/Common/Constants.cs
@@ -16,4 +16,16 @@ public static class Constants
     {
         public const string GeneralMeter = "MinimalTemplate";
     }
+
+    public static class Roles
+    {
+        public const string StandardUserRole = "standard-user-role";
+        public const string AdminUserRole = "admin-user-role";
+    }
+
+    public static class Policies
+    {
+        public const string StandardUser = "StandardUser";
+        public const string AdminUser = "AdminUser";
+    }
 }
