+++ tests/MinimalApiTemplate.Api.Tests/MinimalApiTemplate.Api.Tests.csproj
@@ -18,9 +18,9 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
+    <PackageReference Include="AwesomeAssertions" Version="8.0.2" />
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
-    <PackageReference Include="FluentAssertions" Version="6.12.2" />
     <PackageReference Include="NBuilder" Version="6.1.0" />
     <PackageReference Include="xunit.assert" Version="2.9.3" />
     <PackageReference Include="NSubstitute" Version="5.3.0" />
